third_party/blink/web_tests/svg/custom/SVGException-expected.txt

Tests the properties of the exception thrown by rotateFromVector.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.toString() is "InvalidAccessError: Failed to execute 'rotateFromVector' on 'SVGMatrix': Arguments cannot be zero."
PASS Object.prototype.toString.call(e) is "[object DOMException]"
FAIL Object.prototype.toString.call(e.__proto__) should be [object DOMExceptionPrototype]. Was [object DOMException].
PASS e.constructor.toString() is "function DOMException() { [native code] }"
PASS e.constructor is window.DOMException
PASS successfullyParsed is true

TEST COMPLETE
